week1/article/serializers.py
from article.models import Article, Company
from rest_framework import serializers


class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = ('id', 'name')


class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = ('id', 'title', 'rating', 'summary', 'ip_address',
                  'submission_date', 'company')
