week1/requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.3
django-filter==2.1.0
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.8
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
